feat(tcb):add the listener and remote attribute in the tcb

+++ include/core/stack.h
@@ -14,7 +14,7 @@ namespace myu {
         TimerManager timer_manager_;
         UdpDriver udp_driver_;
     public:
-        TcpStack(const char *listen_ip, int listen_port)
+        TcpStack(const char *listen_ip, uint16_t listen_port)
             : loop_(new uv_loop_t), timer_manager_(loop_), udp_driver_(loop_, listen_ip, listen_port) {
             uv_loop_init(loop_);
         }
